+++ __test__/unittest/template_funcs.test.ts
@@ -1,14 +1,15 @@
-const nunjucks = require("nunjucks");
-const nunjucksConfig = require("../src/builder/nunjucks_config");
+import nunjucks from "nunjucks";
+import nunjucksConfig from "../../src/builder/nunjucks_config";
 const env = nunjucks.configure({ autoescape: false });
 nunjucksConfig(env);
 
 describe("test templateFuncs", () => {
-
   test("substr behavior", () => {
     const str = "abcdefghi";
-    const run = (str, begin, end) => nunjucks.renderString(`{{ test | substr(${begin}, ${end}) }}`, {test: str});
-    let res;
+    function run(str: string, begin?: number, end?: number): string {
+      return nunjucks.renderString(`{{ test | substr(${begin}, ${end}) }}`, { test: str });
+    }
+    let res: string;
     res = run(str);
     expect(res).toEqual(str);
     res = run(str, 1);
@@ -22,12 +23,12 @@ describe("test templateFuncs", () => {
   test("addPrefix behavior", () => {
     const has = "PREFIX:usefulid";
     const hasnot = "usefulid";
-    const run = (str, prefix, delim) => {
-      prefix = typeof prefix === 'undefined' ? undefined : `"${prefix}"`;
-      delim = typeof delim === 'undefined' ? undefined : `"${delim}"`;
-      return nunjucks.renderString(`{{ test | addPrefix(${prefix}, ${delim}) }}`, {test: str})
-    };
-    let res;
+    function run(str: string, prefix?: string, delim?: string): string {
+      prefix = typeof prefix === "undefined" ? undefined : `"${prefix}"`;
+      delim = typeof delim === "undefined" ? undefined : `"${delim}"`;
+      return nunjucks.renderString(`{{ test | addPrefix(${prefix}, ${delim}) }}`, { test: str });
+    }
+    let res: string;
     res = run(hasnot, "PREFIX");
     expect(res).toEqual(has);
     res = run(hasnot, "PREFIX", "-");
@@ -37,11 +38,11 @@ describe("test templateFuncs", () => {
   test("rmPrefix behavior", () => {
     const has = "PREFIX:usefulid";
     const hasnot = "usefulid";
-    const run = (str, delim) => {
-      delim = typeof delim === 'undefined' ? undefined : `"${delim}"`;
-      return nunjucks.renderString(`{{ test | rmPrefix(${delim}) }}`, {test: str})
-    };
-    let res;
+    function run(str: string, delim?: string) {
+      delim = typeof delim === "undefined" ? undefined : `"${delim}"`;
+      return nunjucks.renderString(`{{ test | rmPrefix(${delim}) }}`, { test: str });
+    }
+    let res: string;
     res = run(has);
     expect(res).toEqual(hasnot);
     res = run(has, "-");
@@ -53,12 +54,12 @@ describe("test templateFuncs", () => {
   test("replPrefix behavior", () => {
     const has = "PREFIX:usefulid";
     const hasnot = "usefulid";
-    const run = (str, prefix, delim) => {
-      prefix = typeof prefix === 'undefined' ? undefined : `"${prefix}"`;
-      delim = typeof delim === 'undefined' ? undefined : `"${delim}"`;
-      return nunjucks.renderString(`{{ test | replPrefix(${prefix}, ${delim}) }}`, {test: str})
-    };
-    let res;
+    function run(str: string, prefix?: string, delim?: string) {
+      prefix = typeof prefix === "undefined" ? undefined : `"${prefix}"`;
+      delim = typeof delim === "undefined" ? undefined : `"${delim}"`;
+      return nunjucks.renderString(`{{ test | replPrefix(${prefix}, ${delim}) }}`, { test: str });
+    }
+    let res: string;
     res = run(has, "NEWFIX");
     expect(res).toEqual("NEWFIX:" + hasnot);
     res = run(hasnot, "NEWFIX");
@@ -71,31 +72,32 @@ describe("test templateFuncs", () => {
 
   test("array usage", () => {
     const has = ["PREFIX:id0", "PREFIX:id1", "PREFIX:id2"];
-    const hasnot = "usefulid";
-    let res;
-    let run = (str, delim) => {
+    let res: string;
+    function run1(str: string[], delim?: string, prefix?: string) {
       prefix = typeof prefix === "undefined" ? undefined : `"${prefix}"`;
       delim = typeof delim === "undefined" ? undefined : `"${delim}"`;
       return nunjucks.renderString(`{{ test | rmPrefix(${prefix}, ${delim}) | join(",") }}`, { test: str });
-    };
-    res = run(has);
+    }
+    res = run1(has);
     expect(res).toEqual("id0,id1,id2");
-    run = (str, prefix, delim) => {
+    function run2(str: string[], prefix?: string, delim?: string): string {
       prefix = typeof prefix === "undefined" ? undefined : `"${prefix}"`;
       delim = typeof delim === "undefined" ? undefined : `"${delim}"`;
       return nunjucks.renderString(`{{ test | addPrefix(${prefix}, ${delim}) | join(",") }}`, { test: str });
-    };
-    res = run(has, "test", ":");
-    expect(res).toEqual('test:PREFIX:id0,test:PREFIX:id1,test:PREFIX:id2');
-    run = (str, prefix, delim) => {
+    }
+    res = run2(has, "test", ":");
+    expect(res).toEqual("test:PREFIX:id0,test:PREFIX:id1,test:PREFIX:id2");
+    function run3(str: string[], prefix?: string, delim?: string) {
       prefix = typeof prefix === "undefined" ? undefined : `"${prefix}"`;
       delim = typeof delim === "undefined" ? undefined : `"${delim}"`;
       return nunjucks.renderString(`{{ test | replPrefix(${prefix}, ${delim}) | join(",") }}`, { test: str });
-    };
-    res = run(has, "test");
+    }
+    res = run3(has, "test");
     expect(res).toEqual("test:id0,test:id1,test:id2");
-    run = run = (str, begin, end) => nunjucks.renderString(`{{ test | substr(${begin}, ${end}) }}`, { test: str });
-    res = run(has, -3);
+    function run4(str: string[], begin: number, end?: number) {
+      return nunjucks.renderString(`{{ test | substr(${begin}, ${end}) }}`, { test: str });
+    }
+    res = run4(has, -3);
     expect(res).toEqual("id0,id1,id2");
-  })
+  });
 });
