+++ jest.config.js
@@ -1,3 +1,5 @@
 module.exports = {
-    setupFilesAfterEnv: ['./jest.setup.js'],
+  preset: "ts-jest",
+  testEnvironment: "node",
+  setupFilesAfterEnv: ["./jest.setup.js"],
 };
