+++ package.json
@@ -40,9 +40,12 @@
     "@biothings-explorer/smartapi-kg": "file:../smartapi-kg",
     "@commitlint/cli": "^11.0.0",
     "@commitlint/config-conventional": "^11.0.0",
+    "@types/jest": "^29.5.5",
+    "@types/nunjucks": "^3.2.4",
     "coveralls": "^3.1.0",
-    "jest": "^26.6.3",
-    "standard-version": "^9.1.1"
+    "jest": "^29.0.0",
+    "standard-version": "^9.1.1",
+    "ts-jest": "^29.1.1",
     "typescript": "^5.2.2"
   },
   "husky": {
